Fix a performance bug due to re-reading configuration settings once per output line. This was having a considerable impact on both cpu time and memory. 

The slowness was most notable when processing large piped input (e.g. > 1000 lines). For single lines, the problem was never really noticed.

Now, the configuration settings are cached, and therefore only processed once improving the speed considerably (about 15 times faster!!!)

The easiest way to see the difference is by using the `seq` command to generate 10000 lines, and pipe it to the `c8y util show` utility which just transforms the standard input to the output engine. Below show the results before and after the optimization.

**Before (4.347 seconds)**

```sh
$ seq -f '{"id":"%g"}' 1 10000 | time c8y util show --select id
c8y util show --select id  5.28s user 0.33s system 129% cpu 4.347 total
```

**After (0.313 seconds)**

```sh
$ seq -f '{"id":"%g"}' 1 10000 | time c8y util show --select id
c8y util show --select id  0.20s user 0.05s system 78% cpu 0.313 total
```